# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FinanceDOM API — Subventions non remboursables Outre-Mer
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/dossiers:
    get:
      operationId: listAdminDossiers
      tags: [admin]
      parameters:
        - name: statut
          in: query
          schema:
            type: string
        - name: territoire
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All dossiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDossier"
  /admin/dossiers/{id}:
    get:
      operationId: getAdminDossier
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dossier detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDossier"
  /admin/dossiers/{id}/status:
    put:
      operationId: updateDossierStatus
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDossier"
  /admin/frais:
    get:
      operationId: listAdminFrais
      tags: [admin]
      responses:
        "200":
          description: All fees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Frais"
    post:
      operationId: emitFrais
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmitFraisBody"
      responses:
        "201":
          description: Created fee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Frais"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        prenom:
          type: string
        nom:
          type: string
        email:
          type: string
        telephone:
          type: string
        territoire:
          type: string
        typePorteur:
          type: string
        organisation:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
      required: [id, prenom, nom, email, territoire, typePorteur, role, createdAt]

    # ── Dashboard ──
...
    Dossier:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        titre:
          type: string
        territoire:
          type: string
        dispositif:
          type: string
        secteur:
          type: string
        statut:
          type: string
        montantDemande:
          type: number
        montantApport:
          type: number
        progressionEtape:
          type: integer
        totalEtapes:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, reference, titre, territoire, dispositif, secteur, statut, montantDemande, progressionEtape, totalEtapes, createdAt, updatedAt]
...
    Frais:
      type: object
      properties:
        id:
          type: integer
        dossierId:
          type: integer
        reference:
          type: string
        montantHT:
          type: number
        montantTVA:
          type: number
        montantTTC:
          type: number
        statut:
          type: string
          enum: [en_attente, paye, annule]
        echeance:
          type: string
        paidAt:
          type: string
        createdAt:
          type: string
      required: [id, dossierId, reference, montantHT, montantTVA, montantTTC, statut, echeance, createdAt]
...
    FraisLigne:
      type: object
      properties:
        label:
          type: string
        description:
          type: string
        montantHT:
          type: number
      required: [label, description, montantHT]
...
    AdminDossier:
      allOf:
        - $ref: "#/components/schemas/Dossier"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            expertDesigne:
              type: string
            frais:
              type: array
              items:
                $ref: "#/components/schemas/Frais"
    UpdateStatusBody:
      type: object
      properties:
        statut:
          type: string
        expertDesigne:
          type: string
        commentaire:
          type: string
      required: [statut]
    EmitFraisBody:
      type: object
      properties:
        dossierId:
          type: integer
        lignes:
          type: array
          items:
            $ref: "#/components/schemas/FraisLigne"
        echeanceDays:
          type: integer
      required: [dossierId, lignes]